+++ app/src/groovy/de/laser/api/v1/ApiReader.groovy
@@ -320,6 +320,24 @@ class ApiReader {
         return ApiReaderHelper.cleanUp(result, true, true)
     }
 
+    static exportCostItems(def ignoreRelation, Org context, Date startDate, Date endDate){
+        def result = []
+        if(startDate && endDate) {
+            def costItems = CostItem.findAllByOwnerAndDateCreatedBetween(context, startDate, endDate)
+            costItems.each {
+                result << ApiReader.exportCostItem(it, context)
+            }
+
+        }else{
+            def costItems = CostItem.findAllByOwner(context)
+            costItems.each {
+                result << ApiReader.exportCostItem(it, context)
+            }
+
+        }
+
+        return ApiReaderHelper.cleanUp(result, true, true)
+    }
 
     // ################### HELPER ###################